pyproject.toml
[project]
name = "byd-seriescode"
version = "0.1.0"
description = ""
authors = [
    {name = "Renan Domingos",email = "rna.domingos@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
